third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/responseURL-expected.txt

Test XMLHttpRequest.responseURL

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS responseURL is ""
PASS responseURL is ""
xhr.readyState = 2: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/reply2.txt
xhr.readyState = 3: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/reply2.txt
xhr.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/reply2.txt
PASS status is "200"
PASS responseURL is "http://127.0.0.1:8000/xmlhttprequest/resources/reply2.txt"
PASS responseURL is ""
PASS responseURL is ""
xhr.readyState = 2: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/reply2.xml
xhr.readyState = 3: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/reply2.xml
xhr.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/reply2.xml
PASS status is "200"
PASS responseURL is "http://127.0.0.1:8000/xmlhttprequest/resources/reply2.xml"
PASS responseURL is ""
PASS responseURL is ""
xhr.readyState = 2: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/navigation-target.html
xhr.readyState = 3: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/navigation-target.html
xhr.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/navigation-target.html
PASS status is "200"
PASS responseURL is "http://127.0.0.1:8000/xmlhttprequest/resources/navigation-target.html"
PASS responseURL is ""
PASS responseURL is ""
xhr.readyState = 2: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/not-found.txt
xhr.readyState = 3: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/not-found.txt
xhr.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/not-found.txt
PASS status is "404"
PASS responseURL is "http://127.0.0.1:8000/xmlhttprequest/resources/not-found.txt"
xhr.readyState = 1: responseURL = 
PASS successfullyParsed is true

TEST COMPLETE
